+++ lib/KR.mli
@@ -77,11 +77,17 @@ module Id : sig
   val compare : t -> t -> int
 end
 
-type warning =
-  | Objective_not_found of Work.t
-  | Migration of { work_item : Work.t; objective : Work.t option }
-      (** For retro-compatibility only.
-          This case should be removed once everything has migrated to objectives. *)
+module Warning : sig
+  type t =
+    | Objective_not_found of Work.t
+    | Migration of { work_item : Work.t; objective : Work.t option }
+        (** For retro-compatibility only.
+            This case should be removed once everything has migrated to objectives. *)
+
+  val pp : t Fmt.t
+  val pp_short : t Fmt.t
+  val greppable : t -> string option
+end
 
 val v :
   kind:Kind.t ->
@@ -94,7 +100,7 @@ val v :
 val dump : t Fmt.t
 val merge : t -> t -> t
 val compare : t -> t -> int
-val update_from_master_db : t -> Masterdb.t -> t * warning option
+val update_from_master_db : t -> Masterdb.t -> t * Warning.t option
 
 (** {2 Pretty-print} *)