+++ test/cram/cat/migrate.t
@@ -93,10 +93,9 @@ This weekly is using objectives:
 Linting of the original file fails because we used workitems
 
   $ okra lint -e -C admin eng1.workitems.md
-  [ERROR(S)]: eng1.workitems.md
-  
-  Invalid objective:
-    "Property-Based Testing for Multicore (#1090)" is a work-item, you should use its parent objective "Property-Based Testing for Multicore (#558)" instead
+  File "eng1.workitems.md", line 3:
+  Error: Invalid objective:
+         "Property-Based Testing for Multicore (#1090)" is a work-item. You should use its parent objective "Property-Based Testing for Multicore (#558)" instead.
   [1]
 
 We rewrite the file using okra cat