+++ test/cram/lint/includes.t
@@ -21,9 +21,8 @@ Missing include sections are errors, for engineer reports this is "Last Week"
   > 
   > More unformatted text.
   > EOF
-  [ERROR(S)]: <stdin>
-  
-  Missing includes section: LAST WEEK
+  File "<stdin>", line 1:
+  Error: Missing includes section: LAST WEEK
   [1]
 
 The include_section checker looks for all passed sections
@@ -46,8 +45,6 @@ The include_section checker looks for all passed sections
   > 
   > More unformatted text.
   > EOF
-  [ERROR(S)]: <stdin>
-  
-  Missing includes section: NEXT WEEK,
-  LAST WEEK
+  File "<stdin>", line 1:
+  Error: Missing includes section: NEXT WEEK, LAST WEEK
   [1]