+++ test/cram/lint/migrate.t
@@ -27,6 +27,10 @@ and the error message points to the corresponding objective.
   $ cat > weekly.md << EOF
   > # Last week
   > 
+  > - This objective does not exist (#32)
+  >   - @eng1 (0 day)
+  >   - xxx
+  > 
   > - JSOO Effect Performance (No KR)
   >   - @eng1 (1 day)
   >   - xxx
@@ -52,12 +56,21 @@ and the error message points to the corresponding objective.
   okra: [WARNING] KR ID updated from "No KR" to "#701":
   - "JSOO Effect Performance"
   - "JSOO Effect Performance"
-  [ERROR(S)]: weekly.md
-  
-  Invalid objective:
-    "General okra maintenance (#1115)" is a work-item, you should use an objective instead
-  [ERROR(S)]: weekly.md
-  
-  Invalid objective:
-    "Property-Based Testing for Multicore (#1090)" is a work-item, you should use its parent objective "Property-Based Testing for Multicore (#558)" instead
+  File "weekly.md", line 3:
+  Error: Invalid objective: "This objective does not exist"
+  File "weekly.md", line 19:
+  Error: Invalid objective:
+         "General okra maintenance (#1115)" is a work-item. You should use an objective instead.
+  File "weekly.md", line 15:
+  Error: Invalid objective:
+         "Property-Based Testing for Multicore (#1090)" is a work-item. You should use its parent objective "Property-Based Testing for Multicore (#558)" instead.
+  [1]
+
+  $ okra lint -e -C admin weekly.md --short
+  okra: [WARNING] KR ID updated from "No KR" to "#701":
+  - "JSOO Effect Performance"
+  - "JSOO Effect Performance"
+  weekly.md:3: Invalid objective: "This objective does not exist" (not found)
+  weekly.md:19: Invalid objective: "General okra maintenance (#1115)" (work-item)
+  weekly.md:15: Invalid objective: "Property-Based Testing for Multicore (#1090)" (work-item), use "Property-Based Testing for Multicore (#558)" instead
   [1]